Bug 1823953 - Loop over avcodec_receive_frame when decoding using FFmpegAudioDecoder. r=alwu,media-playback-reviewers DONTBUILD
It seems to work without, but the officiel ffmpeg command line utility and the
example for their internal library all do this.

This also splits the big DoDecode function into three sub-function for clarity,
as it was a bit hard with the #if statements.

Differential Revision: https://phabricator.services.mozilla.com/D173460

UltraBlame original commit: f7f4a7585ceb9406f541d79a5c8dc4eea5dc5695