Bug 1826382 - Teach the WAV demuxer to report the channel layout of WAV files and simplify it. r=alwu DONTBUILD
This remove the FmtParser that was doing very little, and make it so the
FmtChunk report all attribute that it contains, not a subset of them. Depending
on the format, the format chunk can be of difference size. It's frequently 40
bytes when it's a WAVEFORMATEXTENSIBLE.

Differential Revision: https://phabricator.services.mozilla.com/D175156

UltraBlame original commit: 2811d12803cf23549d48e96dce8bfbb01ffd8329