Bug 1826382 - Teach the WAV demuxer to report the channel layout of WAV files and simplify it. r=alwu
This remove the FmtParser that was doing very little, and make it so the
FmtChunk report all attribute that it contains, not a subset of them. Depending
on the format, the format chunk can be of difference size. It's frequently 40
bytes when it's a WAVEFORMATEXTENSIBLE.

Differential Revision: https://phabricator.services.mozilla.com/D175156

UltraBlame original commit: ee63d7b81520a41f1ca3bcf750740af642404b50