Bug 1847196 - Remove python 2 compatibility code from mozbuild.frontend.reader. r=firefox-build-system-reviewers,ahochheiden
The starting point was a Deprecation warning about visit_Str. It turns
out the function used to have an enlightening comment before bug
1616920, which read:
  # String nodes we got from the AST parser are str, but we want
  # unicode literals everywhere, so transform them.

In other words, the sole purpose of that function was to make the
transformed code closer to python 3 even when running in python 2. It's
a no-op on python 3, and, well we don't support python 2 anymore. So the
method can just go away.

While here, it turned out that some of the code that was added in bug 1616920
can also be undone because we don't need python 2 compatibility anymore.
And while we're doing that, let's just remove the uses of six as well.

Differential Revision: https://phabricator.services.mozilla.com/D185421

UltraBlame original commit: 96daa719f4fcaf24978fb15fdc57b8e7299290d7