Bug 1846630 [wpt PR 41269] - [SDA] Fix crash when resolving inset with an unattached subject, a=testonly
Automatic update from web-platform-tests
[SDA] Fix crash when resolving inset with an unattached subject

When resolving a length from an inset, we use a length resolver, which
in turn requires a reference to the subject's style. When the subject
is not attached to the document, its style is not calculated and we
had a null deref. To avoid this problem, we verify that the subject
has a style before using the length converter. We handle the trivial
cases of a literal px or percent inset, and otherwise fall back to
marking the inset as being style dependent, in which case, it needs to
be resolved on each style update.

Bug: 1428965
Change-Id: I870d20a72d6dfd9cb787132022545442150d1ecb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4738874
Reviewed-by: Mustaq Ahmed <mustaqchromium.org>
Commit-Queue: Kevin Ellis <keverschromium.org>
Cr-Commit-Position: refs/heads/main{#1178978}

--

wpt-commits: 611e18d18cc233b5e033a2034cb8b117b02dd685
wpt-pr: 41269

UltraBlame original commit: 5f273ba25ccda096c7f615be2c0f35b2de911676