Bug 1824051 - Fix alias sets for CheckIsObj and CheckScriptedProxyGetResult r=jandem
So, this patch is very tentative. The reason for this is that if we just change
the alias sets for these ops, we end up failing this assertion: https://searchfox.org/mozilla-central/rev/85269d4444c2553e7f4c669fe4de72d64f4fe438/js/src/jit/CodeGenerator.cpp#332.
Our solution to this is to give the ops resume points, and... blow right past
the warnings telling us not to have multiple effectful operations for a single
transpiled IC. Iain suggested that I rope you in here, Jan, to get your take on
whether this is safe or not. Nothing seems to blow up in tests.

Differential Revision: https://phabricator.services.mozilla.com/D185350

UltraBlame original commit: ca712509847a99ccb801642467dcf5f08ec1186b