Bug 1749766 - fix test_logoutAndTeardown.js by having it send a long response r=jschanck
Previously the simulated server in test_logoutAndTeardown.js would send a short
response, which meant that the client could receive and process the response
before necko closed all connections in reaction to the call to
logoutAndTeardown.
This patch changes the server to send an indefinitely long response, which
means that the only way forward is for the client to properly close all
connections (or time out or run out of memory).

Differential Revision: https://phabricator.services.mozilla.com/D186022

UltraBlame original commit: 51c1600372ae1fc4b34d3f3d0d77d993ebba6dee