Bug 1724048 - Only add the HttpChannelChild to the loadGroup when guaranteed we will actually open the channel. r=necko-reviewers,kershaw
When we AsyncOpen the HttpChannelChild, we might exit early, possibly because
we're already shutting down. But if we've already added the channel to the
loadgroup when the loadgroup gets cleaned up we try to cancel the channel.
Because we didn't get to that point where we call SetEventTarget,
we would trigger the neckoTarget assertion in HttpChannelChild::AsyncCallImpl.

This patch makes sure we add the channel to the loadgroup when we are certain
it will actually be opened.

Differential Revision: https://phabricator.services.mozilla.com/D186123

UltraBlame original commit: 0e5fd894164e17697e806038e640065b22a274a7