Bug 1848090 - When parsing JSON, Stop allocating in the nursery after two nursery collections r=sfink
The patch adds a nursery collection callback for the duration of parsing and
counts the collections. After the second collection we switch to allocating
everything in the tenured heap.

This change made parsing the JSON data in the bug significantly faster.

Differential Revision: https://phabricator.services.mozilla.com/D186362

UltraBlame original commit: fc9820aab42fd045da58360329c9514e860e0da0