Bug 1841097 - Implement a checkbox in ETP Custom subpanel for the configuration of the fingerprinting protection. r=anti-tracking-reviewers,fluent-reviewers,flod,bvandersloot,mstriemer
This patch implements a checkbox in ETP Custom usbpanel. It's for the
configuration of the fingerprinting protection. Similar to the tracking
protection checkbox, it has a dropdown list to select which window to
enable the protection, private windows or all windows.

In addition, we update the existing copy for fingerprinters to reflect
the newly added fingerprinting protectoin checkbox.

Differential Revision: https://phabricator.services.mozilla.com/D186023

UltraBlame original commit: 24109a1476e63f026968a0b8ca434a646b61de7b