Bug 1170986 - patch 0 - Reformat gfxMacPlatformFontList.mm with clang-format as though it were a .cpp file. r=gfx-reviewers,nical
No functional change. The issue here is that we're going to move a lot of the contents of this file
to a new .cpp source file; but we have different line-length specifications for .mm vs .cpp files,
and so that results in clang-format making a lot of whitespace-only changes to adjust the lines.

To try and simplify review, this preparatory patch is created by running clang-format on the file
*as if it were a .cpp source*, so that the whitespace adjustments get done. Then the following patch,
which moves much of the content to CoreTextFontList.cpp, will already see the whitespace in the form
it expects.

Differential Revision: https://phabricator.services.mozilla.com/D180324

UltraBlame original commit: 31bf68247e6e3ba7f84ca7dc493bf5d9aa0167ee