Bug 1850434 - part2 : add crash test. r=media-playback-reviewers,azebrowski
The test file is from [1], which is under GNU Free License, so we can
use it for crash test.

This file has samples which are over the metadata duration, and the last
sample has invalid negative duration.

[1] https://en.wikipedia.org/wiki/File:Close-mid_central_rounded_vowel.ogg

Depends on D187499

Differential Revision: https://phabricator.services.mozilla.com/D187500

UltraBlame original commit: 6dd8caac7e278326fbda1597a6c9845863314ba8