Bug 1848193 - Check H264 data when converting from AVCC to AnnexB r=media-playback-reviewers,alwu
Instead of ignoring the conversion error, `ConvertAVCCTo4BytesAVCC`
should only return `Ok` when the given AVCC data can be converted.

When converting H264 data in AVCC format into AnnexB format, the first
step is to convert the AVCC data into 4-byte-header AVCC data. In this
step, `ConvertAVCCTo4BytesAVCC` ignores the error if it fails to read
the NAL body in the length that header claims to be. In order to check
if the given data is corrupt, it needs to return an error in this case.
Without returning an error, it can surprisingly return a `Ok` when no
conversion is done, if the first NAL header gives an incorrect NAL body
length.

In addition, to check if 4-byte AVCC data is corrupt, we should not
early return a `Ok` when the extra data claims the length is 4-byte. We
still need to read the NAL body to ensure the body length matches the
length claimed in the NAL header.

Depends on D185971

Differential Revision: https://phabricator.services.mozilla.com/D185972

UltraBlame original commit: 65f68d60351198047933d799f7777913c20289e4