Bug 1851976 - Mark infallible unbox non-movable. r=iain
We were adding an infallible unbox for the array passed to `JSOp::SpreadCall`.
This is fine, but we then loop hoisted it before the branch checking if the value
is an object or `undefined` (from `JSOp::OptimizeSpreadCall`). The unbox then failed
for the case where the value is `undefined`.

Differential Revision: https://phabricator.services.mozilla.com/D187709

UltraBlame original commit: 8de24ac4d2477fd592c3ff1619f44260d9b28629