Bug 1852101 - Always destroy out-of-flow frames from the placeholder. r=dholbert,layout-reviewers
This is more consistent, and allows me to remove the destruct root from
the DestroyContext.

Try is green with this, reasons for it being conditional are probably
historical.

This is in preparation for not needing a destruct root in DestroyContext,
and plumbing a DestroyContext consistently.

Differential Revision: https://phabricator.services.mozilla.com/D187692

UltraBlame original commit: 53c402edec263b039de12dac3363021745937579