Bug 1760747. Ensure that the bounds on the geometry of filter items are always limited by the building rect. r=jrmuizel
There are several paths in ComputeGeometryChange where we can allocate an geometry item for a filter item, but only one path that will call DetectContainerLayerPropertiesBoundsChange (which adjusts the bounds for filter items). Make sure this always happens in ComputeGeometryChange.

This was causing a bug where the geometry bounds were the full filter item height on the first paint (even though only part of the filter item was visible inside the building rect), but we did not call DetectContainerLayerPropertiesBoundsChange because it was the first paint where the filter item was visible. Then on the next paint the full filter item was visible, and there is no rect change to signal that we need to repaint.

Differential Revision: https://phabricator.services.mozilla.com/D187549

UltraBlame original commit: 12882fe96164b756c60fff4549300977b85af06f