Bug 1749504 - Convert TestQuotaManager.cpp from a file containing generic quota tests to a file dedicated for QuotaManager testing only; r=dom-storage-reviewers,jstutte
New tests need to be added for the QuotaManager class, but TestQuotaManager.cpp
currently contains tests which are not directly related to the QuotaManager. It
would be better to keep them in separate files.

Changes done in this patch:
- OriginScope related tests moved to a new file TestOriginScope.cpp
- WARN_IF_FILE_IS_UNKNOWN tests moved to TestQuotaCommon.cpp
- A simple test for QuotaManager::ShutdownStorage added to TestQuotaManager.cpp

Differential Revision: https://phabricator.services.mozilla.com/D185524

UltraBlame original commit: 3d179c2f922d5dc24ce016b1ebd8cd32e1bbe121