Bug 1749504 - Add QuotaManager::IsShuttingDown check to Quota::RecvClearStoragesForPrivateBrowsing; r=dom-storage-reviewers,jstutte
The clearing operation used to use a sub actor, so the IsShuttingDown check was
done in Quota::AllocPQuotaRequestParent. Bug 1840772 stopped using a sub actor, but a corresponding IsShuttingDown check was not added to a new method for
handling the clearing operation.

Changes done in this patch:
- AssertIsOnOwningThread added to QuotaManager::ClearPrivateRepository
- IsShuttingDown check added to Quota::RecvClearStoragesForPrivateBrowsing

Differential Revision: https://phabricator.services.mozilla.com/D185527

UltraBlame original commit: b9b7f97dfc95255f97438bfd88617e1611084485