Bug 1851785 [wpt PR 41839] - De-flake getFrameStats WPTs., a=testonly
Automatic update from web-platform-tests
De-flake getFrameStats WPTs.

The discardedFrames was incorrectly assuming that no frames would be
discarded initially, but the test only uses ideal so frames can be
discarded, and our implementation may sometimes discard the first frame
so this assertion is removed.

Also timeouts have been observed on the bots. The bot output[1] does
not say which of the tests are timing out but the most likely candidate
is the test that waits for a calculated timeout - which could be large
if the bot is under heavy load. This test is updated to wait for a
fixed number of milliseconds instead.

Lastly, for good measure, since all of these tests wait for media to
flow, the timeout should be "long".

[1] https://ci.chromium.org/ui/p/chromium/builders/ci/Mac12%20Tests/12050/test-results?sortby=&groupby=

Bug: chromium:1472978
Change-Id: If6ddb8f78409a2bec08c2abb8ab28baccdc75454
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4844860
Commit-Queue: Henrik Boström <hboschromium.org>
Reviewed-by: Tony Herre <topricechromium.org>
Cr-Commit-Position: refs/heads/main{#1192947}

--

wpt-commits: 92cf2872002be4c0da58bbeb928a42350c4c2169
wpt-pr: 41839

UltraBlame original commit: 997baf7032df156281dda89318a2f23d39a26c3e