Bug 1851735 - Adding accessible name to the chevron button on Shopping cards. r=shopping-reviewers,jhirsch
While the button is included in the `<summary>` element, it is still can be accessed in screen reader's browse mode or with the mouse tracking (by hovering over the button with NVDA, for example). We need to make sure the button is clearly labelled and does communicate its expanded/collapsed state programmatically, since we're exposing this control on-screen.

Differential Revision: https://phabricator.services.mozilla.com/D187806

UltraBlame original commit: 4d8c9e2984f48092117faa1faf2c6811afd34260