Bug 1853467 part 1 - Add fast path to Object.assign for when we can reuse the shape's PropMap. r=jonco
We already had an optimization to reuse the shape when assigning to an empty object.
This often doesn't work because the number of fixed slots is different, but in
this case we can still reuse the property map.

Differential Revision: https://phabricator.services.mozilla.com/D188966

UltraBlame original commit: bb511dab103a75d21f9d1a0171f897c881dc0928