Bug 1834274 - Part 1: Implement nsIRFPService.idl for nsRFPService. r=emilio
To allow remote settings client to set fingerprinting overrides from
Javascript, we need to implement nsIRFPService for nsRFPService.

This patch changes the return format of nsRFPService::GetOrCreate() to
match the singleton implementation of nsIRFPService. As a result of it,
we need to change the first init call in nsContentUtils because we
cannot forward a already_AddRefed<> to mozilla::Unused. It will cause a
memory leak.

Differential Revision: https://phabricator.services.mozilla.com/D182809

UltraBlame original commit: 79b200c1060e3eba94dea597da093617ec8fc868