Bug 1855174 - Process xpcshell node output in realtime r=jmaher,necko-reviewers,kershaw
Spawned nodejs servers sometimes print output to stdout or stderr.
While it's not strictly necessary to display, it's sometimes useful for
debugging, but it's problematic that we only dump the output when killing
the node server.

After landing D189212 this becomes even more problematic, as killing the
process via os.killpg means the file descriptors are already closed.

This patch makes it so for each started server we also spawn a new thread
that reads from stdout and stderr. This way the output is present in the
logs at the correct time.

Depends on D189212

Differential Revision: https://phabricator.services.mozilla.com/D189213

UltraBlame original commit: 45571d960e8a9df0e42fd383075781b2a90b6e07