Bug 1838393 - [devtools] Tweak paused thread styling. r=devtools-reviewers,bomsy.
Instead of only displaying a small icon when the thread is
paused, add a plain text paused label.
When a thread is paused and not selected, set its background
and color the same as we have for the why paused element.

CSS variables for paused background and color are created and
are used in both WhyPaused and paused thread.

We take this opportunity to refactor Threads.css a bit, cleaning
up spacing and using CSS nesting to make it easier to read.

Differential Revision: https://phabricator.services.mozilla.com/D189341

UltraBlame original commit: d17297618424c47052180c28bcfb4d6cc19af316