Bug 1855680 - Add fast path for global regexp replace without capture groups. r=iain
Use the mechanism added in bug 1839422 to add a fast path for global regexp replace with
a callback function. In this case we can avoid allocating the result object and also optimize
the call to the callback function.

We could also use this optimization for the `ELEMBASE` optimization, but it's less common
(we don't hit it at all on Speedometer 3) and it requires duplicating more code.

Differential Revision: https://phabricator.services.mozilla.com/D189487

UltraBlame original commit: 7813c893a1f8dca29947cdfb26c56e40e4a862d1