Bug 1839396 part 3 - Extract Sprinter::putAsciiPrintable with mozilla::Span. r=mgaudet
QuoteString is currently specialized with a Sprinter argument, but to do that it
relies on the `operator[]`, which implies being able to read the bytes. While
there is no issue today, this goes against making the *Printers infallible.

Thus, this change moves the append logic to a dedicated `putAsciiPrintable`,
removing the `getOffset()` and `operator[]` usage from `QuoteString`.

Differential Revision: https://phabricator.services.mozilla.com/D181488

UltraBlame original commit: 129aeb743d8ef3f923436ca776fb013f4ef4cabf