Bug 1835263 - Fix tabswitch timeout. r=Gijs,perftest-reviewers,AlexandruIonescu
In bug 1848357 I simplified the setup of the MozLayerTreeReady events in
a way such that if we already have layers we don't get such event
repeatedly every time.

This test was listening for that event, and thus timing out causing
various crashes. This adapts the test in a similar way to how the
tab switcher got adapted in bug 1851486.

I couldn't repro the timeout locally but the start of these timeouts
matches the landing of bug 1848357, so I'm fairly certain this should
help.

Differential Revision: https://phabricator.services.mozilla.com/D189635

UltraBlame original commit: 3677fd4aebcb91190a1d79fc4deeee12bf903671