Bug 1845407 - Don't pump apple events before application starts. r=nika
When we call `nextEventMatchingMask` we send a "did launch app" notification to the OS workspace. If we do that before we call `[NSApp run]` accessibility isn't hooked up yet internally in AppKit and we fail to respond to a11y API queries.

Pumping Apple events is currently used for getting url requests in order to launch Firefox with the URLs loading correctly in a browser window. We can achieve the same result by using `application:openURLs:` in the delegate and calling `run` on the NSApp earlier and then stopping the loop in order to build the command line and finish initializing Firefox.

Also pumping the apple event queue should be deferred in `ProcessNextNativeEvent` until `[NSApp run]` is called.

Differential Revision: https://phabricator.services.mozilla.com/D184524

UltraBlame original commit: c869f080aad6301b8c8021152a97de1e23de2f24