Bug 1854658 - Fix the part combinator to deal with exportparts correctly. r=zrhoffman
We need to jump to an element in the correct scope just like ::slotted()
does.

Factor that code to its own function both so that it's easier to reason
about and so that the code for that function remains small. At the end
of the day other combinators like descendant or sibling get executed a
ton of times (while pseudo-elements only jump once).

Differential Revision: https://phabricator.services.mozilla.com/D189478

UltraBlame original commit: d4e50ba4e3bb32cd117ee7c751c056f37b00e6f4