Bug 1846474 - Part 2: add `enum class TrapMachineInsn`.  r=rhunt.
This adds `enum class TrapMachineInsn` and a couple of utility routines for
it.  A `TrapMachineInsn` indicates roughly what kind of machine instruction we
expect to see at a trap site (read/write+size, or UD, or "Atomic") as a basis
for later (post-compilation) comparison.  The new enum is not used for
anything in this patch.

Differential Revision: https://phabricator.services.mozilla.com/D187457

UltraBlame original commit: 8b4b42a31cd242b712f0da907d062a1d996a32fc