Bug 1856547 - Remove revert-llvmorg-14-init-11890-gf86deb18cab6.patch. r=sergesanspaille
The patch it reverted was causing problems while running tests with
ASan, but, from a try push, it seems those are not happening anymore.
Presumably, the issue was either fixed on the workers (whatever that
might have been), or in newer versions of clang.
Either way, we're not using clang 14 for ASan builds either, so we can
remove the patch from clang 14 too.

Differential Revision: https://phabricator.services.mozilla.com/D189908

UltraBlame original commit: 01abf930014033b2a323982947a0027d248b0a4f