Bug 1856487 [wpt PR 42217] - Paint background invalidation when deleting the last element of TableRow, a=testonly
Automatic update from web-platform-tests
Paint background invalidation when deleting the last element of TableRow

When the last element in a TableRow is deleted, it does not trigger
a redraw of the background,causing the background to persist.
After removing the last element of a row in a Table,
the size of the TableRow remains unchanged because it does not affect
the layout of the preceding elements.
Therefore, caching is used during rendering.
We should proactively trigger a redraw of this row when deleting
the last element of a table row.

Bug: 1487554
Change-Id: Ia45f5abf8ccd54781e8f0c1de04aebfa5cad552c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4900123
Reviewed-by: Rune Lillesveen <futharkchromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhuchromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhuchromium.org>
Cr-Commit-Position: refs/heads/main{#1204290}

--

wpt-commits: 6447f83eae831c2470b4e8eb5cd6af6fbbe0103e
wpt-pr: 42217

UltraBlame original commit: 4529ae5e3a5ee1be5ed6f25ab23a04f0d4211285