Bug 1816998: Replace macOS quarantine database-based attribution with extended attribute attribution r=nalexander
With extended attribute attribution using exactly the same attribution strings as Windows (eg: no URLs), this turns out to be quite straightforward. Once we've pulled in the attribution string from the extended attribute, the existing parsing, validation, etc. just works.

The only wrinkle is that the extended attributes may have nul bytes or tabs that we need to strip away. (Tabs may be present because we use them to pad the attribution area when we prepare the DMG for attribution. Nul bytes may be present because we overwrite the entire attribution before updating the attribution data.)

Differential Revision: https://phabricator.services.mozilla.com/D189258

UltraBlame original commit: c937eafeffe2965d2e1dd3523bfb7f235429da73