Bug 1571426 - Read geckodriver port number in fixture, r=webdriver-reviewers,jdescottes
Instead of passing in a port selected in the fixture, pass in 0 as the
port, allowing geckodriver to select the port, and then read the port
number from the utput in stdout.

This should prevent us trying to reuse an already bound port in
a race-free way.

Differential Revision: https://phabricator.services.mozilla.com/D189926

UltraBlame original commit: 4e85807cccd784e48f93eba438bb756a9ff3990c