Bug 1854636 - Temporarily disable some Tier 2 a11y_checks for Settings UI. r=Jamie,Gijs,settings-reviewers
We are working on [enabling Tier 2 a11y-checks to ensure our products have basic accessibility built-in (bug 1692110)](https://bugzilla.mozilla.org/show_bug.cgi?id=1692110) and before that, we need to prepare the existing code base. While we triage and investigate existent tests that would fail once the a11y-checks are enabled in the CI, we are going to skip the failing tests altogether and then handle each one of them individually.

This task is to skip with `fail-if` condition of the failing tests while we continue investigation into the reasons these tests failed and to be able to backtrack these tests once they're resolved.

Related bugs:
- Disabling tests: meta bug 1848402
- Investigation and remediation for components with disabled tests: meta bugs 1848394 and 1854227, Settings UI-specific meta bug 1854637
- Explore capturing click events for a11y_checks: bug 1692110

Differential Revision: https://phabricator.services.mozilla.com/D188991

UltraBlame original commit: 36460a147684643e3c1b2b6b17055192e8f59487