Bug 1851056 - Fix assert when GMPContentParent::SendChromiumCDMConstructor fails. r=media-playback-reviewers,padenot
If SendChromiumCDMConstructor fails, then the diagnostic assert in
ChromiumCDMDestroyed will fail because we have yet to add the
ChromiumCDMParent to the array. Now in this patch we just add the parent
right away, and rely upon the ActorDestroy machinery to remove it as in
all other cases.

Differential Revision: https://phabricator.services.mozilla.com/D188963

UltraBlame original commit: d57c8b63407ee78e1d5aa4a9e3e93f620b3407ba