Bug 1845932 - Rework mach-try-perf commit message to be shorter. r=perftest-reviewers,AlexandruIonescu DONTBUILD
This patch changes the mach-try-perf commit message (first line) to be much shorter, and limited to 200 characters. Instead of displaying both the queries, and the actual selections on the same line, this patch splits it into two lines. This provides better formatting options for the queries displayed as well.

Differential Revision: https://phabricator.services.mozilla.com/D190090

UltraBlame original commit: 60af720985092b2072845a378ab68a4c3f98670a